Safari-611.1.10/LayoutTests/media/modern-media-controls/localized-strings/replaced-string-expected.txt

Testing that we can replace strings using UIString.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS UIString('Test replaced string', 15) is "Value 15 should be printed here"
PASS successfullyParsed is true

TEST COMPLETE
